add auth and review microservice implementations

+++ app/entity/review.go
@@ -2,6 +2,6 @@ package entity
 
 type Review struct {
 	ID      uint64
-	Mark    uint8
+	Mark    uint32
 	Comment string
 }
